api/requirements.txt
fastapi
uvicorn
pillow
sqlalchemy
pymysql  # MySQL을 사용하는 경우
